Move Slide background next to Theme + add group separators
Slide background is a document-level control; it now lives in
RightGlobals alongside the Theme toggle so it stays reachable
from every toolbar state (idle / object / text-edit), not only
when nothing is selected.

Adds separators at every group boundary that was previously
butted up against the next group:
- contextual middle ↔ right globals
- Done ↔ slide-style group (text-edit state)
- right globals ↔ Present split-button
- shape-controls: Fill ↔ Border
- text-element-controls: Border ↔ Font
- image-controls: Reset crop ↔ Alt text

IdleSection collapses to just InsertGroup since the Background
button moved out. The background-handler test is renamed to
reflect its new home in RightGlobals (logic is unchanged).
